Add files via upload
This table defines the geological "land units" of the study area. Each land unit is comprised of one or more of the geological units identified in the model "Reconstruction South Eat Asia, 0-20My" (Melia, UP 2019).

The composition of the land unit is adjusted as needed per 1My timeslice over the 20My study period to reflect the functional "one-ness" of the unit at any given time.